sdk/js: refactor nft test utils
